♻️ refactor: 매뉴얼 오버레이를 별도 컴포넌트로 분리
매뉴얼 오버레이 UI와 로직을 main.basic.js에서 ManualOverlay.js 컴포넌트로 추출
